pyproject.toml
[project]
name = "langchain-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-openai>=1.1.3",
    "langchain-surrealdb>=0.2.0",
    "numpy>=2.4.2",
    "python-dotenv>=1.2.1",
    "surrealdb>=1.0.8",
]
